Refactor: restructure userstats module
