+++ .gitignore
@@ -11,7 +11,7 @@ client_secret.json
 *.log.gz
 *.bak
 temp.csv.gz
-userstats/
+/userstats/
 
 # Covers JetBrains IDEs: IntelliJ, RubyMine, PhpStorm, AppCode, PyCharm, CLion, Android Studio and Webstorm
 # Reference: https://intellij-support.jetbrains.com/hc/en-us/articles/206544839
