+++ kaztron/cog/userstats/__init__.py
@@ -1,0 +1,5 @@
+from kaztron.cog.userstats.userstats import UserStats
+
+
+def setup(bot):
+    bot.add_cog(UserStats(bot))
